+++ pycfmodel/model/resources/generic_resource.py
@@ -3,6 +3,7 @@ from typing import ClassVar
 
 from pydantic import Extra, validator
 
+from pycfmodel.model.generic import Generic
 from pycfmodel.model.resources.resource import Resource
 from pycfmodel.model.resources.types import ResourceModels
 
@@ -12,10 +13,11 @@ _EXISTING_RESOURCE_TYPES = {klass.TYPE_VALUE for klass in ResourceModels.__args_
 
 
 class GenericResource(Resource):
-    """This class is used for all resource types that we haven't had time to implement yet"""
+    """This class is used for all resource types that don't have a dedicated class."""
 
     ALLOW_EXISTING_TYPES: ClassVar[bool] = True
     Type: str
+    Properties: Generic
 
     class Config(Resource.Config):
         extra = Extra.allow